Database Update
Updated database and relivent backend methods to support canvas integration

+++ student-5/backend/GradesManager/GradesManager/DTOs/StudentAssignmentDto.cs
@@ -4,7 +4,7 @@
     (
         Guid StudentId,
         Guid AssignmentId,
-        int? TempMark,
-        int? FinalMark
+        double? TempMark,
+        double? FinalMark
     );
 }
